requirements.txt
pyicloud>=2.0.0
PyYAML>=6.0
flask>=3.0
